UPDATE: replace AuditLogArchiverService with AuditLogService and remove unused archiver service
